Update types Create an APReportingUnit type; this is the way race results will be delivered. (In our case, we'll only have one reporting unit per race, as we want statewide results only). This contains data on precincts reporting as well as the candidates.
Update the APRace interface to use the APReportingUnit one, as well as make optional a few fields that do not always come through in API responses and are not necessary for our purposes.

Add an optional 'runoff' property to our Candidate interface; the AP's API reports whether a candidate has won a race outright or is advancing to a runoff.